Prototype landing page of a real estate company
This is my first repository project and its a landing page of a Real Estate company. I put my courage into testing, Then my html structure couple with my cascading stylesheet. In the coming project I'll be able to complete a full website responsively for mobile and iPad. But for now rate my first unresponsive project as a newbies in frontend development with HTML & CSS aspiring to become a Full-stack developer in coming months.